Add devDependency and loader util for additional bundler
This adds `@rspack/core` as a `devDependency` to Next.js, as it will be conditionally referenced in core in the future and will only run when the real dependency is added via a future plugin.

This also creates a `getRspackCore` utility which performs this check, suppresses the lint, and more clearly informs the user of the error.